@@ -1,35 +1,54 @@
 package com.example.asset.config;
-
-import com.example.asset.security.JwtAuthFilter;
-import com.example.asset.security.JwtVerifier;
-import org.springframework.beans.factory.annotation.Value;
+import com.example.common.security.JwtAuthFilter;
+import com.example.common.security.JwtVerifier;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 
+/**
+ * ✅ Security configuration for Notification Service.
+ * Uses shared JwtVerifier (from common module) to validate tokens
+ * using either RSA public key or HMAC secret.
+ */
 @Configuration
 public class SecurityConfig {
 
-    @Value("${security.jwt.public-key-path:classpath:keys/jwt-public.pem}")
-    private String publicKeyPath;
+    private final JwtVerifier jwtVerifier;
 
-    @Bean
-    public JwtVerifier jwtVerifier() {
-        return new JwtVerifier(publicKeyPath);
+    public SecurityConfig(JwtVerifier jwtVerifier) {
+        this.jwtVerifier = jwtVerifier;
     }
 
     @Bean
-    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
-        JwtAuthFilter jwtFilter = new JwtAuthFilter(jwtVerifier());
-        http.csrf(csrf -> csrf.disable())
+    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
+        // Custom JWT validation filter
+        JwtAuthFilter jwtAuthFilter = new JwtAuthFilter(jwtVerifier);
+
+        http
+            .csrf(csrf -> csrf.disable())
             .authorizeHttpRequests(auth -> auth
-                .requestMatchers("/actuator/**", "/public/**").permitAll()
-                .requestMatchers("/api/**").authenticated()
-                .anyRequest().denyAll()
+                // Asset API endpoints require authentication
+                .requestMatchers("/api/asset/**").authenticated()
+                // Permit actuator/health or any public endpoints if needed
+                .requestMatchers("/actuator/**").permitAll()
+                // Swagger/OpenAPI endpoints
+                .requestMatchers(
+                    "/swagger-ui/**",
+                    "/swagger-ui.html",
+                    "/v3/api-docs/**",
+                    "/api-docs/**",
+                    "/swagger-resources/**",
+                    "/webjars/**"
+                ).permitAll()
+                .anyRequest().permitAll()
             )
-            .addFilterBefore(jwtFilter, UsernamePasswordAuthenticationFilter.class);
+
+            .addFilterBefore(jwtAuthFilter, UsernamePasswordAuthenticationFilter.class);
+
         return http.build();
     }
 }
+
+
