@@ -1,5 +1,4 @@
 
-
 package com.example.authservice.repository;
 
 import com.example.authservice.model.User;
@@ -11,27 +10,41 @@ import java.util.Optional;
 @Repository
 public interface UserRepository extends JpaRepository<User, Long> {
 
-    // ✅ Existence checks using compositeId fields (Spring Data will auto-resolve)
+    // Simple existence checks
     boolean existsByCompositeId_UsernameHash(String usernameHash);
     boolean existsByCompositeId_EmailHash(String emailHash);
     boolean existsByCompositeId_MobileHash(String mobileHash);
 
-    // ✅ Standard finders for deterministic hash lookups
+    // Single field finders
     Optional<User> findByCompositeId_UsernameHash(String usernameHash);
     Optional<User> findByCompositeId_EmailHash(String emailHash);
     Optional<User> findByCompositeId_MobileHash(String mobileHash);
 
-    // ✅ Lookup by numeric user ID (native PK)
+    // Primary key lookup
     Optional<User> findByUserId(Long userId);
 
-    // ✅ Multi-tenant lookup: username + projectType
+    // Multi-tenant lookups
     Optional<User> findByCompositeId_UsernameHashAndCompositeId_ProjectType(String usernameHash, String projectType);
-
-    // ✅ Multi-tenant lookup: email + projectType
     Optional<User> findByCompositeId_EmailHashAndCompositeId_ProjectType(String emailHash, String projectType);
-
-    // ✅ Multi-tenant lookup: mobile + projectType
     Optional<User> findByCompositeId_MobileHashAndCompositeId_ProjectType(String mobileHash, String projectType);
+
+    // ✅ Corrected: composite lookups (username + mobile/email + projectType)
+    Optional<User> findByCompositeId_UsernameHashAndCompositeId_MobileHashAndCompositeId_ProjectType(
+            String usernameHash, String mobileHash, String projectType);
+
+    Optional<User> findByCompositeId_UsernameHashAndCompositeId_EmailHashAndCompositeId_ProjectType(
+            String usernameHash, String emailHash, String projectType);
+
+    // Existence checks per tenant
+    boolean existsByCompositeId_UsernameHashAndCompositeId_ProjectType(String usernameHash, String projectType);
+    boolean existsByCompositeId_EmailHashAndCompositeId_ProjectType(String emailHash, String projectType);
+    boolean existsByCompositeId_MobileHashAndCompositeId_ProjectType(String mobileHash, String projectType);
+    
+    boolean existsByCompositeId_UsernameHashAndCompositeId_MobileHashAndCompositeId_ProjectType(String usernameHash,String mobileHash, String projectType);
+    boolean existsByCompositeId_UsernameHashAndCompositeId_EmailHashAndCompositeId_ProjectType(String usernameHash,String mobileHash, String projectType);
 }
+
+
+
 
 
