@@ -1,31 +1,54 @@
+
 package com.example.notification.config;
 
-import com.example.notification.security.JwtAuthFilter;
-import com.example.notification.security.JwtVerifier;
-import org.springframework.beans.factory.annotation.Value;
+import com.example.common.security.JwtAuthFilter;
+import com.example.common.security.JwtVerifier;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 
+/**
+ * ✅ Security configuration for Notification Service.
+ * Uses shared JwtVerifier (from common module) to validate tokens
+ * using either RSA public key or HMAC secret.
+ */
 @Configuration
 public class SecurityConfig {
 
-    @Value("${jwt.public-key-path}")
-    private String publicKeyPath;
+    private final JwtVerifier jwtVerifier;
 
-    @Bean
-    public JwtVerifier jwtVerifier() {
-        return new JwtVerifier(publicKeyPath);
+    public SecurityConfig(JwtVerifier jwtVerifier) {
+        this.jwtVerifier = jwtVerifier;
     }
 
     @Bean
-    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
-        JwtAuthFilter filter = new JwtAuthFilter(jwtVerifier());
-        http.csrf(csrf -> csrf.disable())
-            .authorizeHttpRequests(auth -> auth.anyRequest().authenticated())
-            .addFilterBefore(filter, UsernamePasswordAuthenticationFilter.class);
+    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
+        // Custom JWT validation filter
+        JwtAuthFilter jwtAuthFilter = new JwtAuthFilter(jwtVerifier);
+
+        http
+            .csrf(csrf -> csrf.disable()) // Disable CSRF for REST APIs
+            .authorizeHttpRequests(auth -> auth
+                // Swagger/OpenAPI endpoints
+                .requestMatchers(
+                        "/swagger-ui/**",
+                        "/swagger-ui.html",
+                        "/v3/api-docs/**",
+                        "/api-docs/**",
+                        "/swagger-resources/**",
+                        "/webjars/**"
+                ).permitAll()
+                // Only allow POST to /api/notifications with valid JWT
+                .requestMatchers("/api/notifications/**").authenticated()
+                // Permit actuator/health or any public endpoints if needed
+                .requestMatchers("/actuator/**").permitAll()
+                .anyRequest().permitAll()
+            )
+            .addFilterBefore(jwtAuthFilter, UsernamePasswordAuthenticationFilter.class);
+
         return http.build();
     }
 }
+
